Swap ModInfos._mergeable_by_type for ModInfo.merge_types:
WIP merge_types will become a method - note:
- I allow for any combination of types - this seems to be the case
already although as of now MERGE and ESL/OVERLAY do not appear in any
game together - but see #596
- fileInfo is already in scope most of the time (while modInfos always)
- we only use the whole cache in checkMods - rest are lookups. This was
a hint that the cached checks belong to the info in this case - this
parallels the work in #336 and #353
- I had to pass the game_handle to isPBashMergeable in place of the
unused _mod_header_reader - I moved the static functions from the
latter to ModInfo - they belong there plus this simplified the
lambdas in error_msgs.

Mopy/bash/basher/__init__.py: the  `+ ' '` had to go

Don't use merge_types rather use cached_types in expensive/do once
operations.

In these operations the performance of creating the 'mergeable' list
does not matter. Turns out that merge_types duplicates the mergeInfo info
- we should keep the latter only, but that makes is_mergeable slow
enough to be better to iterate any way. Note a subtlety in
plugin_types._depended: we are actively scanning for mergeability so
potentially the merged info of dependents is not updated - in
_refreshMergeable it *is* already updated due to the reverse=True in:

	for fn_mod, modInfo in dict_sort(self, reverse=True,
									 key_f=load_order.cached_lo_index):

(except for delinquent master?) - which we should repeat in
refreshMergeable? See also bc1d02a909434192f03470879df2d6b17db57961.

I had introduced an is_mergeable Mod/MasterInfo pair but reverted -
the current handling is a bit clumsy but that is really cause
MERGE is still to find its place - revisit as part of #596.
Note that merge_types becomes such a pair further down the branch,
wrapping the clumsy mergeInfo property.