Nuke overrides of set_mod_flag:
That was a big remaining rough edge - note mod_info.get_extension() and
self._mod_info_attr uses went down - and we only setattr for the latter
in (now final) set_mod_flag. It would be more natural to call
update_onam from a MasterFlag override but I prefer it final - note
that set_mod_flag does not return these weird internal bools (used in
the overrides). I also absorbed one use in set_plugin_flags - the
remaining two directly set tes4.flags1 and it is difficult to absorb
(edit: this led to a few bugs - I ended up removing the check_mod_flags
call from set_plugin_flags, and adding it to the only place it might be
needed, in mod_links the link would not be enabled if the mod already
had a conflicting flag).