FFF don't use merge_types rather use the cached ones in expensive operations.
In these operations the performance of creating the 'mergeable' list
does not matter. Turns out that merge_types duplicates the mergeInfo info
- we should keep the latter only, but that makes is_mergeable slow
enough to be better to iterate any way. There is one exception that performance
seems to matter - plugin_types._depended - there is a subtlety here:
we are actively scanning for mergeability so potentially the merged
info of dependents is not updated - in _refreshMergeable it *is* already
updated due to the reverse=True in:

        for fn_mod, modInfo in dict_sort(self, reverse=True,
                                         key_f=load_order.cached_lo_index):

(except for delinquent master?) - which we should repeat in
refreshMergeable ? EEE TTT

RRR see "Simplify rescanMergeable return value"

I had introduced an is_merheable Mod/MasterInfo pair but reverted -
the current handling is a bit clumsy but that is really cause
MERGE is still to find its place - revisit as part of # <--- RRR  596