FFF don't use merge_types rather use the cached ones in expensive operations.
In these operations the performance of creating the 'mergeable' list
does not matter. Turns out that merge_types duplicates the mergeInfo info
- we should keep the latter only, but that makes is_mergeable slow
enough to be better to iterate any way. There is one exception that performance
seems to matter - plugin_types._depended - and this reveals a subtlety
This uses the is_mergeable check while scanning for mergeability -
potentially the merged info of dependents is not updated - in _refreshMergeable
it is already updated due to the reverse=True in:

        for fn_mod, modInfo in dict_sort(self, reverse=True,
                                         key_f=load_order.cached_lo_index):

which we should repeat in refreshMergeable ? EEE TTT

squash! Simplify rescanMergeable return value: TTT b32121b1bb2b4ee15e0ad972cbce78538c6656a2