Refactor _mastersStart handling - re-add master block size test to SF:
Observe that _mastersStart is used:

- in SaveFileHeader._do_write_header where it used to dump ins till
that point
- in _load_masters_16 where it is used for the size test

In the second case we add the sse_offset then subtract it - in the
first case the offset is added only if the file is compressed but then
in _do_write_header we have special code in this case that does not use
_mastersStart. So seems we can drop sse_offset and use a
plugin_info_size attribute to do the master size check. Now that the
dust has settled we could use some more refactoring in save headers -
note that we can define position marker attributes by passing
ins.tell() as the lambda. Maybe we could use those to skip the master
block on dumping instead of rereading them (see for instance
SaveFileHeader._dump_masters, where we could
`ins.seek(self.plugin_info_size, 1)`). We could also skip some
attributes or del _mastersStart for instance to make sure it won't be
misused.
I dropped the ins param from SaveFileHeader init as it seems unused -
and we need to trace ins creation - seems to be a _io.BytesIO or
BufferedReader instance.