[!]Make sure mergeInfo is always updated on reverting:
Still muddy water here - one of the last few pools that stand in drains.
Making set_plugin_flags update mergeability seems the correct thing -
but do we need to _update_info_sets first? Stashing a [!] but as stated
below we must first finish up #353.

Sort input for rescanMergeable:

Seems the right thing also - anyway we are about to finalize the
core refresh APIs so this can be revisited after the dust settles
and we actually have centralized refresh operations.