WIP nit in save_headers:
I removed the TODO on BSA files as from a cursory look we are doing
different things there - still collecting the functions in the enum
is a step towards generalizing this - maybe a better place would be
archives.py. For now this is too saves specific though.

Nit in set_item_format - the return value was not used since
65c74c8055df59fb3d096635d4e29d07682cc8c7.