balt.Installer_Op: RRR
Centralize BAIN refresh logic - 7 RefreshUI calls (and issue_warnings)
dropped. ui_refresh_ in the UI and refresh_ui_ in BAIN are reserved
for passing along function parameters. refresh_ui and ui_refresh are
the variables to track - the goal here is we pass more detailed info
to the panels RefreshUI.
Note Installer_Anneal Execute was not decorated with conversation - if
wasn't needed won't harm to add it, if it was then this is a bugfix.
Same goes double with issue_warnings, more of whose were missing. Then
issue_warnings had one use and turns out the final use of ui_refresh
variable so no harm adding the data_store_key to it at this point. WIP!

Under # 353 core - see next commits