propagate_refresh:
We can afford a new API in exchange for the refresh_others param. This
will permit us to pass kwargs for the RefreshUI arguments making
refresh more granular. WIP - in particular we could consider adding
distribute_warnings - but I absolutely like the simplification of
RefreshUI, both its signature and that we stop nesting RUI calls in RUI
itself - note the ugly deprint gone.

Note:

- distribute_ui_refresh does not require a defaultdict anymore as
it iterates on ui_refresh.items()
- saves_links.py: the one case where focus_list is False - maybe we
need more of it?
- DeleteItems on saves would issue the warning in RefreshUI - TTT duh
- ui_refresh[self.window.data_store_key] = True was not really needed
as we are called from iPanel which is not used in distribute_warnings