WIP _bain_op TTTT
The mirror of _Installers_Op - albeit much more complex.

Here are the changes in the signatures - the ui_refresh_ were for tracking
the use sites of ui_refresh - not needed anymore due to passing those
as kwargs

FFF WIP _bain_op TTTTTTT EEE split notify_external changes?? split DO

The goal here was two fold - in balt we created ui_refresh as a
defaultdict(bool) - so calling RefreshUI on all data store. But then
on _fs_install we were issuing all the store refreshes *per install*,
so I would have to centralize this. Were the refreshes used in the next
install op? This needed reviewing core BAIN and resulted in this mini
branch.

- RefrIn objects learned how to `|` themselves - RefrIn is used
as the input of refresh so we need to keep updating it, for operations
that need to redraw the items (like ceding ownership to another installer)
or for deletion updates.

Note we perform deletes during the operation and we refresh on the spot -
we don't want that -> do_refr param

The challenge here is to make sure all accesses to the _bain_op
decorated methods are through _perform_action

The code added to RefrData.__ior__ is just for the
Installers_MonitorExternalInstallation special case.

Absorb progress to kwargs also

SSS Cleanup notify_external

Nit - remove DO()

Store is now ready to be used as the glue between bosh globals and the
UI

FFF bain_clean_data_dir

FFF fixups for monitor external last=True?

Correct handling of data_sizeCrcDate: SSS FFF TTT

_fs_install: We must do this immediately after having installed the
files as data_sizeCrcDate is used in next operations. That is what the
code was doing, but in order to get the ftime from the store we had
to refresh the data stores *per* _fs_install call (note also the ugly
access of .corrupted attribute if refresh failed). TTT as the updates
of crc I do are untested - note we would like to pass the cached_stat
to the infos

SSS drop the os error:

@@ -1364,3 +1364,2 @@ def _fs_install(self, dest_src, srcDirJoin, progress, subprogressPlus,
         for dest, (s, c, dest_path) in data_sizeCrcDate_update.items():
-            try:
             try:
@@ -1372,7 +1371,2 @@ def _fs_install(self, dest_src, srcDirJoin, progress, subprogressPlus,
                 d = dest_path.mtime
-            except OSError: ##: too much?
-                idata_data_scd.pop(dest, None)
-                if st_fn := dest_to_store.get(dest):
-                    rui_data[st_fn[0]].new_or_present.pop(st_fn[1], None)
-                continue
             idata_data_scd[dest] = (s, c, d)

Note:

__init__.py  233 __init__: Failed to set k='crc' to v=454675967 for self=BSAInfo<Maskar's Oblivion Overhaul - Meshes.bsa>
__init__.py  233 __init__: Failed to set k='crc_size' to v=847955927 for self=BSAInfo<Maskar's Oblivion Overhaul - Meshes.bsa>
__init__.py  233 __init__: Failed to set k='crc_mtime' to v=1566934805.9670966 for self=BSAInfo<Maskar's Oblivion Overhaul - Meshes.bsa>
__init__.py  233 __init__: Failed to set k='crc' to v=1799247724 for self=BSAInfo<Maskar's Oblivion Overhaul - Sounds.bsa>
__init__.py  233 __init__: Failed to set k='crc_size' to v=21848244 for self=BSAInfo<Maskar's Oblivion Overhaul - Sounds.bsa>
__init__.py  233 __init__: Failed to set k='crc_mtime' to v=1566934805.9979968 for self=BSAInfo<Maskar's Oblivion Overhaul - Sounds.bsa>
__init__.py  233 __init__: Failed to set k='crc' to v=3264193280 for self=BSAInfo<Maskar's Oblivion Overhaul - Textures.bsa>
__init__.py  233 __init__: Failed to set k='crc_size' to v=1647927505 for self=BSAInfo<Maskar's Oblivion Overhaul - Textures.bsa>
__init__.py  233 __init__: Failed to set k='crc_mtime' to v=1566934870.764402 for self=BSAInfo<Maskar's Oblivion Overhaul - Textures.bsa>

WIP fix wizards - split out install/anneal logic

Wiz fixups TTT FFF

The issue was calling a _bain_op on a row - not possible. Note this means
we used to refresh per package installed. Not sure why we do this
anneal vs install - could we use install all along?

FFF create the edits before the bain_op to refresh