[!]WIP _bain_op:
The mirror of _Installers_Op - albeit much more complex.

The goal here was twofold - in balt we created ui_refresh as a
defaultdict(bool) - triggering a full RefreshUI on affected data stores.
But then on _fs_install we were issuing all the store refreshes *per
install*, so I would have to centralize this. Were the refreshes used
in the next install op? This needed reviewing core BAIN and resulted in
this mini branch commit.

- RefrIn objects learned how to __ior__ themselves - RefrIn is used
as the input of refresh so we need to keep updating it, for operations
that need to redraw the items (like ceding ownership to another
installer) or for deletion updates.

The code added to RefrData.__ior__ is just for the
Installers_MonitorExternalInstallation special case.

Note we perform deletes during the operation and we refresh on the spot,
we don't want that -> do_refr param - note the deletions were happening
early in _remove_restore while now we perform them after the
_remove_restore call - should be ok still [!]

All accesses to the _bain_op decorated methods must go through
_perform_action

Correct handling of data_sizeCrcDate:

_fs_install: We must do this immediately after having installed the
files as data_sizeCrcDate is used in next operations. That is what the
code was doing, but in order to get the ftime from the store we had
to refresh the data stores *per* _fs_install call (note also the ugly
access of .corrupted attribute if refresh failed). [!] as the updates
of crc I do are untested - note we would like to pass the cached_stat
to the infos - we would need to expose _stat_tuple then.

Drop the OSError:

@@ -1364,3 +1364,2 @@ def _fs_install(self, dest_src, srcDirJoin, progress, subprogressPlus,
         for dest, (s, c, dest_path) in data_sizeCrcDate_update.items():
-            try:
             try:
@@ -1372,7 +1371,2 @@ def _fs_install(self, dest_src, srcDirJoin, progress, subprogressPlus,
                 d = dest_path.mtime
-            except OSError: ##: too much?
-                idata_data_scd.pop(dest, None)
-                if st_fn := dest_to_store.get(dest):
-                    rui_data[st_fn[0]].new_or_present.pop(st_fn[1], None)
-                continue
             idata_data_scd[dest] = (s, c, d)

Note I had to locally import modInfos to get rid of warnings on bsas:

__init__.py  233 __init__: Failed to set k='crc' to v=454675967 for self=BSAInfo<Maskar's Oblivion Overhaul - Meshes.bsa>
__init__.py  233 __init__: Failed to set k='crc_size' to v=847955927 for self=BSAInfo<Maskar's Oblivion Overhaul - Meshes.bsa>
__init__.py  233 __init__: Failed to set k='crc_mtime' to v=1566934805.9670966 for self=BSAInfo<Maskar's Oblivion Overhaul - Meshes.bsa>
....

WIP fix wizards - split out install/anneal logic:

The issue was calling a _bain_op on a row - not possible. Note we used
to refresh per package installed. Not sure why we do this anneal vs
install ([!]) - could we use install all along?

Create the edits before the bain_op to refresh

Under #353 and #219