FFF WIP _bain_op TTTTTTT EEE split notify_external changes?? split DO
The goal here was two fold - in balt we created ui_refresh as a
defaultdict(bool) - so calling RefreshUI on all data store. But then
on _fs_install we were issuing all the store refreshes *per install*,
so I would have to centralize this. Were the refreshes used in the next
install op? This needed reviewing core BAIN and resulted in this mini
branch.

- RefrIn objects learned how to `|` themselves - RefrIn is used
as the input of refresh so we need to keep updating it, for operations
that need to redraw the items (like ceding ownership to another installer)
or for deletion updates.

Note we perform deletes during the operation and we refresh on the spot -
we don't want that -> do_refr param

The challenge here is to make sure all accesses to the _bain_op
decorated methods are through _perform_action

The code added to RefrData.__ior__ is just for the
Installers_MonitorExternalInstallation special case.

Absorb progress to kwargs also

SSS Cleanup notify_external

Nit - remove DO()

Store is now ready to be used as the glue between bosh globals and the
UI

FFF bain_clean_data_dir

FFF fixups for monitor external last=True?