Move load_cache from init to _reset_cache:
Compliments of `load_cache=load_cache` in AFile init. This removed most
dummy uses of load_cache and also setting it explicitly to False. Note
also kwargs accesses - load_cache was expected. Also revealed the subtle
differences with _InstallerPackage.__init__ complete with incomplete docs,
I went for a new param here (fs_load). This is copying/renames territory
(this is the correct name also).

Note do_update path is not affected. This is still WIP see next commits
where load_cache is essentially seen to be mod/saves only.

load_bsa - WIP:

We need to pass load_cache=True via __init__ (or do_update) to init the
self._assets variable. We are falling back to the class _assets variable
but this will change soon when decoupling bsas and mods.
So let's have a new param for load_bsa - be explicit here. We need
to encapsulate __init__ (to the point of a `_in_factory` param in
ListInfo init that will raise if False) but that passes through
abs_path handling - ghost of #336.