_WithMastersInfo:
One of the remaining rough edges from #336 - triggered by #689, as I
needed to be clear about when/how _reset_cache was called.

_WithMastersInfo cleanup:

- reduce uses of _reset_masters (also in order to track that cosave
refesh glitch)
- load_cache absorbed in kwargs - turns out it's mods/saves only
- moved FileInfo.__init__ up
- removed `if load_cache` in INIInfo._reset_cache - see:

https://github.com/wrye-bash/wrye-bash/commit/004c8a7317d8ac143effec660c63a9bd4d33514b#diff-841d4a6baa98922d59ac2a4746b5e559acdfd92321ba93a21b23e910f23882aeR1766

Keep an eye for redundant cache resets - should not happen, as
_reset_cache is only used with __init__ where we should define the
attribute - and do_update where if _reset_cache is called with
(load_cache=False) an exception was raised so we can reset anyway.