Fix for cosave update glitch:
Cosaves were not recognized till a tab out/in - so the second refresh we
did. Probably this broke as back as when refresh was moved to init,
we're talking 5507bbaca87683fc2d942b257e00ae5cc236448e. The cosave_types
were empty - we encounter this elsewhere (like _assets in this branch) -
consider not assigning a default to a class variable (or set it to
None), so we get an attribute error rather than iterate over the empty
default.

Simplify do_update signature:

The other (dark) half of __init__ (~20 uses while __init__ is all but
encapsulated in factory/new_info). Note AFile.do_update absorbs kwargs
already - I like kwargs very much, they are the final solution to the
problem of tracking initialization/use site pairs for variables that
are passed along the parameters of a few methods between those sites,
without being used. The only thing that's missing is documenting those
in the base method - PEP8 stuff (dust *must* settle first).

Under #336 (and #353 as all things do_update)