Fixup in BP activation code:
lo_activate would not expect fileName to be in _active_wip, as the
test check_active_limit([*self._active_wip, fileName]) would be
meaningless and could fail in an edge case (look ma, a fencepost!). So I
went (again) through lo_activate uses, and the only case where an active
mod might be passed in appears to be mod_links.Mod_RebuildPatch.Execute.
Therefore, I adjusted lo_activate to activate masters if an already
active plugin is passed to it, which seems to be the intended behavior.
If this assumption holds, then we probably saved the BP separately to
track activated masters - the new APIs handle this, so I simplified it
to only save once. Finally, I added an output RefrData param to
_execute_bp to collect the modInfos refresh result deep in
PatchExecute, allowing me to prune another propagate_refresh(True).

Minor cleanup in _execute_bp.