load_order.cached_lo_last_esm()
We don't need _lo_wip here - this finalizes the cleanup of ModInfos tier
of load order handling. All ModInfos methods are decorated and ready to
be melded in a variation of lo_de/activate_exact. refreshLoadOrder is
the last piece that needs some work amongst the (3) _lo_cache methods.

May #578 rest easy.