Fixups for updating details when switching back to Bash:
I initially added a use_rdata param to RefreshUI that I needed to
propagate to callers - a mess. Better specify the detail item explicitly
in the few cases we need to do so. Not sure about refreshCurrent
but this should be previous behavior here. Super welcome the
simplification of RefreshUI itself.