ListInfo.named_as to centralize case sensitive comparisons:
Fixing some UX bugs - I moved testChanges up to achieve this and make
sure fileStr is-a str. More work is needed but note we got rid of the
single unique_key override - unique_key > named_as so this is the way
to go - see also comments in this merge commit.