Add support for Oblivion Remastered
Set the saves directory correctly

Un-hardcode the assumption that this dir is always
dirs['saveBase'].join('Saves'), because in this game it's
Saved\SaveGames. Also add support for disabling all save
profiles-related links, because that INI key doesn't work in Oblivion
Remastered.

Get the Saves tab functional

The saves format seems completely new and there does not seem to be an
easy way to read masters or any other data, unfortunately. Manual saves
write some of the data we want to display into the name, but autosaves
do not seem to do that. Still, implemented parsing for it.

Introduce dirs['exe']

To represent the directory where the launch_exe resides (and thus any
programs that want to be placed next to that EXE should go). This fixes
version detection for all the various tools and Oblivion itself.

Note that I got rid of the weird version stripping in app_buttons. It
looks weird with the Oblivion Remastered versions and I never understood
why we do this anyways. Displaying all-zero versions as empty strings
makes sense though, so I kept that.

Update record definitions

Building a BP seems to work now :)

Add support for game INIs above the Data folder

Oblivion Remastered does not have INIs in My Games. While WB has support
for loading the game INI from the app folder, that's Oblivion specific
and relies on an INI key. Additionally, it sets more than just the game
INI path (e.g. it changes the save directory as well, since that INI key
made the game avoid the My Games folder entirely). I preserved support
for that in case Oblivion Remastered still uses that key.

Introduce bass.dirs['lo']

That logic should not be sitting in load_order.py, it's clearly part of
the initialization logic (and game-specific to boot). Introduce a new
dir variable for it and set it correctly to handle Morrowind, the
bUseMyGamesDirectory complexity and now Oblivion Remastered.

And with that, Oblivion RE support should be pretty solid :)
BAIN is still lacking the ability to install PAKs, of course, but that
will be much harder and does not impede an MVP of sorts for Oblivion RE.

Split mods_dir into two variables

That tiny TODO was hiding a real rabbit hole :(
Still, good to get this assumption ripped out as well.

Add support for MGEF codes without terminator

Some plugins for Oblivion Remastered seem to not have this null
terminator, most notably the unofficial patch. Just add it in if it's
missing, the base game files have it for me.

Disable OBSE plugin installation

OBSE plugins have to go into
OblivionRemastered/Binaries/Win64/OBSE/Plugins, which means we need to
be able to install into *three* separate folders. Disable the
installation so people don't wonder why their installed OBSE plugins
don't load.

Write Oblivion.esm to plugins.txt

Turns out Oblivion Remastered crashes if the game master is missing from
plugins.txt.