[!]GPath in canonize_ci_path:
canonize_ci_path is (usually?) passed a path in so it is already GPathed

to_os_path was more difficult to track - there is a single case where
we pass a str (in _KeywordSelect) - we cache those paths now but these
should be few and we do never want to use a bare Path constructor (to
track Path creation for #368). Note that if we don't pass a str here we
could drop the `return _GPath(ci_path)` in the linux implementation.

I went ahead and required to_os_path receiving a Path always -
_image_resource_dir is promoted to a Path. I used to count those `.s`
(#543...) now I don't mind a couple here and there. It's the long
bumpy way towards #368 (actually this has low priority now #543 is
settled).

Chopping off a couple direct Path.__init__ uses - the if check in bash
did not make much sense Path.sys_fs_enc is never Falsy (thanks Infernio
for fixup!) - see:

https://github.com/wrye-bash/wrye-bash/commit/558e4185e022118418ae94db3ab4b7ee5d82c306#r156028348

Finally I dropped os.fspath from the windows implementation - [!] for a
second pair of eyes:

@@ -1249,3 +1249,3 @@ def canonize_ci_path(ci_path: _Path | str) -> _Path | None:
     # Windows is case-insensitive, nothing to do here
-    return _GPath_no_norm(os.fspath(ci_path))
+    return _GPath_no_norm(ci_path)

Makes no sense for path and in all strings come from to_os_path which
got a Path instance for questionable_path argument.

Lastly seems canonize_ci_path never is passed (so returns) None - we
should drop it from the return type if so.