Fixups for error handling in details:
Turns out _Corrupted was failing with a suppressed OSError in case of
ghosts and that exclude was still broken. Note however we don't need
to pass exclude=True here as writeHeader will have recalculated the
excluded attributes (crc etc)

balt.conversation is not needed in _refresh_detail_info as it's only
called in DoSave - the rdata return value will be used presently

Refactor _refresh_detail_info:

Chop off another propagate_refresh=True

I had to corrupt the file while in the debugger to figure all this out -
at least itsa_ghost only now appears in ModInfo(s), not _Corrupted
(thankfully the latter has a single use). Note also that the external
manipulation of RefrData attributes is not needed anymore.

RefrIn is fairly new - we could investigate what happens if we used the
old info in this and other cases we have it available. In this case we
should probably add `rdata.redraw.add(mod_inf.fn_key)` manually as we
do for changeDate - as do_update in refresh will likely return False.

Under #701