Fixup for unhiding installers:
Got a set in _list_store_dir - on patching that I run into the hack of
expecting a list to be projects - this whole hack is traced back to
the three implementations of move_infos - calling refresh three times.
As mentioned in the past, the correct solution is being able to create
infos outside of the store dir - that's some work, so here are some PoC
APIs - I used those to simplify the rename_operation BAIN override, by
not special-casing the marker (so much) and took the opportunity to
chop off another ui_refreshes[...] = True. Last but far from least,
set_path_keys ListInfo API is introduced. This innocuous looking method
will be the end of a few clunky APIs around rename/backup/hide etc.

Under #353/#580