Drop default value for refresh_infos refresh param: RRR
So now we can trace full refreshes by searching for `refresh(True`. Not
all of the (few) occurrences are bad - some are necessary like the one
in RefreshData. Note a couple RefreshUI(True) are now passed a rdata:

@@ -3292,4 +3292,4 @@ def ShowPanel(self, **kwargs):
         """Panel is shown. Update self.data."""
-        if bosh.screen_infos.refresh():
-            self.uiList.RefreshUI(focus_list=False)
+        if rdata := bosh.screen_infos.refresh(True):
+            self.uiList.RefreshUI(rdata, focus_list=False)
         super(ScreensPanel, self).ShowPanel()

Note also I dropped one weird occurrence of refresh:

@@ -3602,3 +3602,2 @@ def warnTooManyModsBsas(self):
                 return # Limit-fixing xSE plugin installed
-        if not len(bosh.bsaInfos): bosh.bsaInfos.refresh()
         infos_num = len(bosh.bsaInfos) + len(bosh.modInfos)

bsaInfos is not None at this point (else len(None) would boom) - so
it is as refreshed as it can be - see init_stores

Note finally that I dropped _boot_refresh_args override for IniInfos -
not needed since 91d05cd9a70d834047e0759da0e727c01f110ee5 as we refresh
the target always but on boot (that is IniInfos.init).

Under # RRR 353