[!]WIP #580/#241/#336:
WIP commit that addresses core issues in rename_operation API. For one
we now call env.shellMove once per rename operation and not per file
(centralizing those ops (and the error checking logic) will come handy
for #241). The for-loop logic is in one place now, at the expense
of an input parameter to rename_operation - this is a refactoring
artifact that will be amended later in the branch. Was trickier than it
seems - in particular markers handling. This sets the ground for the
rest of the work in rename_operation in this branch, including removing
most of its overrides (under the ghost of #336).

Mopy/bash/basher/__init__.py: details rename logic needed a couple
passes to get right - note the peculiar saving logic - we should use
the revert backup logic here and save a known good copy.

bolt:

I had code like:

@@ -1976,4 +1976,3 @@ def __ior__(self, other):
         # we suppose `other` is more up to date so deleted infos might reappear
-        self.to_del -= other.new_changed() | (self.renames.keys(
-            ) & other.renames.values())
+        self.to_del -= other.new_changed()
         self.to_del |= other.to_del

but not really needed as new_changed should contain the reappearing
infos via to_add

get_selected_infos_filtered became one use - UIList namespace can always
use some pruning

[!]try_set_version Do the operations at once:

@@ -3055,4 +3055,4 @@ def try_set_version(self, set_version, *, do_swap=None):
         try:
-            for inf_fname in (baseInfo, move_to), (swapped_inf, master_esm):
-                file_info_rename_op([inf_fname], ren_data, try_once=do_swap)
+            inf_target = [(baseInfo, move_to), (swapped_inf, master_esm)]
+            file_info_rename_op(inf_target, ren_data, try_once=do_swap)
         except CancelError:

envShellMove seems to be able to handle this but does this work
correctly on Linux?[!] - The issue is the renames dict {A:B, C:A} where
A == Oblivion.esm

I had some code as in:

@@ -1982,2 +1982,5 @@ def __ior__(self, other):
             attr.difference_update(other.to_del)
+        for k in self.ren_paths.keys() & other.ren_paths.values():
+            del self.ren_paths[k]
         for att in ('renames', 'ren_paths'):

but bain_notify handles this alright (only use of ren_paths)