Drop _EditableMixinOnFileInfos: RRR
This unfortunate class was deprecated on inception back in
3d36634fd01bfe159006d34c257f480087488463. Have a look to get a feeling
of where we are coming from. There is still some work needed as seen
by the direct calls in superclasses __init__ methods and the boilerplate
code for checking for changes or updating the edit controls - but what
remains is like 0.01% of the effort already put in those details panels.
Note that file_infos needs be overriden per class - this is similar
to`_store` - typing should help here RRR