Refactor irefresh input parameter:
To align with _AfileInfos.refresh. The first good axing of complexity
under #716, absorb InstallersData._list_store_dir to
DataStore.list_store_dir using the machinery from
f41d009124d26cbfa62fcc4a2a15730f17c3f214. This makes update_installers
one-use and permits us to pass an input parameter to the RUI call,
eliminating another full refresh on showing the Installers panel,
(#353). We also removed the omod handling from ShowPanel which is enough
for now - rest of ShowPanel issues are under #701. I remade
list_store_dir private and marked it final.

Note I had to change default status value to None so the return value
of the first refreshStatus includes all packages/markers - happily uses
were few. Markers status is set to -20 - and status handling left for

Two complications that contributed to (commit - author).date = 5 months:

- the handling of the "rename bug" - I ended up using set_path_keys:

@@ -1972,5 +1972,3 @@ def _get_info(self, k, kws, new_or_present):
             deprint(f'{k} invalid idata key: {inst.fn_key}')
-            del self[k] # some rename bug - delete the stored installer
-            # really belongs to refresh_info.redraw but RUI merges with to_add
-            inst = None
+            inst.set_path_keys(k) # rename bug - set paths, rest should be ok
         new_or_present[k] = (inst, kws)

do_update is bypassed on fresh_load but then the info would only be
in this invalid state on fresh load so we don't need to worry about
adding it to rdata, due to status handling described above. Anyway it
should not matter (this was a just a nightly regression, few people
affected). One use of new_changed() and redraw less - good, will get
even better in the last commit on this branch.

- trying to add a progress bar and failing (see
223f5e5afb072b7cc68734103d8e8edd3affdec5, also have a look to see how
all this started). The fix is to show a BusyCursor instead - works like
a charm but this time I will fill an issue -> #728.