[!]Refactor bash tags refresh:
That's not nit anymore - the main achievement here is moving the
responsibility to the info which means among others that mods which have
their tags automatically reloaded will be added to RData.redraw - note
_refresh_bash_tags didn't modify rdata. Architecturally we got rid of
two APIs while the client code was drastically simplified and became
sharper - change checks are centralized and the double anti-pattern of
filtering on get (not on load) and call the setter from the getter is
no more. This was made possible by utilizing the already developed
machinery:

- the processing of persistent attributes on load
- the "Broadcast kwargs" (™) pattern on refresh/init/do_update
- the existing override of do_update

autoBashTags is set on load and tags are reloaded on refresh ->
do_update, removing the weird handling in _refresh_bash_tags and
reducing ModInfos responsibilities to passing the list of files on
refresh - this is still WIP however, as the Bash tags folder is really
super-store responsibility - see #701. This (partly) cleared the
(eternal) fog in restore (follow exclude=True occurrences) - since we
create new infos we should *not* clear these values, as on init setting
auto_tags will correctly update the tags for autoBashTags=True, else
will keep the manually assigned tags, which is what the user most
probably expects (specs missing #292).

Finally this also unearths a bug in rename - we should rename the bash
tags files also - fortunately we now have the tools to handle this.

Three* APIs - the table refactoring APIs are now mature enough to start
using the attributes -> is_auto_tagged is inlined.

Fix for trying to get the description tags while header=None (load_cache
False):

      File "*****\bash\bosh\__init__.py", line 415, in __init__
        self.set_auto_tagged(auto, bt_contents)
      File "*****\bash\bosh\__init__.py", line 710, in set_auto_tagged
        wip_tags = self.getBashTagsDesc()
                   ^^^^^^^^^^^^^^^^^^^^^^
      File "*****\bash\bosh\__init__.py", line 690, in getBashTagsDesc
        if not (re_match := __tags_search(self.header.description)):
                                          ^^^^^^^^^^^^^^^^^^^^^^^
    AttributeError: 'NoneType' object has no attribute 'description'

There are few uses of (modInfos.)factory where load_cache=False and we
never access the bashTags in those cases - still [!] as this is fragile
(but tractable thanks to abfadc944157578fe6ca65196312eabd458bcec8).
Calling do_update should set the tags - I moved it at the top of the
class, as it is an AFile override. We might want to differentiate
between AFile and ListInfo do_update - fs_update for the former? -
one of the main gears driving refresh and we want to have few uses. Note
the complex return logic of set_auto_tagged - only used in do_update for
now, to notify the UI for tags change - we might need to redraw (think
NoMerge/Deactivate). There is still a refresh glitch here that will be
addressed soon (see the OnDismiss comment).

Under #353.