Radiation from _bain_op:
In particular refactor _notify_bain and cleanup notify_external, one
had to wonder every time if the handling of renames here is correct -
so I filled the rest of parameters on site and dropped renames.

Nit - remove DO()

Store is untangled from refresh ready to serve as the glue between bosh
globals and the UI.

Better handling of del_infos:

Only add to to_del items that were actually popped - this is a bugfix
for some edge cases like uninstalling a corrupted package.