Call refresh from rename_operation:
The input type logic in DataStore refresh is used here finally. Chops
off one override of rename_op, complete with super and _notify_bain,
_lo_move_mod became single-use and a refresh call was removed from
Mopy/bash/basher/files_links.py - as seen I encapsulated rename_op in
try_rename (refresh_ui is reincarnated as a bool). Note:

- actually not calling refresh was a bug - for instance the
self.dependents cache was never updated - _file_or_active_updates could
use some cleanup and optimization for renames but laterz
- for unhide, ren_paths are outside the data dir but still it does not
matter yet - IData.track will be revisited and this is super-store
responsibilities anyway, so under #701
- dstore.refresh(rinf, what='I') should pass unlock_lo=True (would warn
on unhiding)
- try_set_version was greatly simplified. Note we are now shamelessly
checking `if master_esm in self` and that voCurrent/_voAvailable are
set in refresh. `act_mods` is a short lived refactoring artifact, I
decided not to activate after renames - see next commit. Error handling
is #241 territory, so we will tackle it after #353 - in particular
rename_op should be used for hiding (in a couple commits) and backup
operations also (tougher).
- details handling needed a couple passes to get right - the end result
is a much simpler _extra_changes implementation but note we need to
explicitly create a RefrData instance in case of a plain redate -
otherwise the _file_or_active_updates is skipped inside refresh, so the
redated attribute not used.